Refactor undo redo
