+++ EasyReflectometryApp/Gui/Pages/Project/SideBarBasic.qml
@@ -28,7 +28,7 @@ EaComponents.SideBarColumn {
                 onClicked: EaGlobals.Variables.showProjectDescriptionDialog = true
                 Component.onCompleted: {
                     ExGlobals.Variables.createProjectButton = this
-                    ExGlobals.Constants.proxy.resetUndoRedoStack()
+                    ExGlobals.Constants.proxy.undoredo.resetUndoRedoStack()
                 }
             }
 
@@ -39,7 +39,7 @@ EaComponents.SideBarColumn {
                 onClicked: ExGlobals.Variables.samplePageEnabled = true
                 Component.onCompleted: {
                     ExGlobals.Variables.continueWithoutProjectButton = this
-                    ExGlobals.Constants.proxy.resetUndoRedoStack()
+                    ExGlobals.Constants.proxy.undoredo.resetUndoRedoStack()
                 }
             }
